Refactor regexp constant
