+++ web/src/const/index.js
@@ -1,1 +1,2 @@
 export * as QUESTION_TYPES from './QuestionTypes';
+export * as REG_EXP from './regExp';
