+++ web/src/const/regExp.js
@@ -1,3 +1,1 @@
-export const REG_EXP = {
-  Number: /^(\s*|\d+)$/,
-};
+export const NUMBER = /^(\s*|\d+)$/
